Added fake answers and answers count  for answers table

+++ database/migrations/2019_10_18_105635_create_answers_table.php
@@ -15,8 +15,8 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('question_id');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('question_id');
+            $table->unsignedInteger('user_id');
             $table->text('body');
             $table->integer('votes_count')->default(0);
             $table->timestamps();
